Make `|` easier to type
Pretty much the only sequence I have to type that I have problems with
is `|>` which I type a lot when I'm programming in elixir.

As it was, I use my thumbs to shift and my thumbs to change layers, so I
was having to move my thumb from key to key to solve this problem.

This change puts them both on the same layer in an easy finger roll,
while staying consistent with the other patterns going on on that layer.