TODO: Lots of useful space
I have a lot of really useful space that I'm just wasting. I'm wondering
what I should do with it.

My hunch is that I probably need to move some of the more useful symbols
over to raise... mostly so that I can stay on home row for as much as
possible. I'm not yet sure exactly what yet though.