Add an option to hold space
Because I use the space hold for layers, I can't hold a key at the
moment. The only place that this really hurts me is gaming (I can't hold
WASD!) or holding spacebar to scroll.

I'm sure there's a setting somewhere to set a max-hold length for a
modtap, but this works too.