Switch my 4x12 layout to the weird miryoku style
Typing on this right now, and it's really weird, but I think I like it.

The main benefit for me is that even on a small grid I still get three
thumb keys, mostly matching my Kyria and the new crkbd for which I
haven't finished the firmware.

This way I'm sharing the three thumb keys macro from thing to thing, but
I am not sure about how I feel about the inner keys.